migrated dists to long

+++ Connect4/src/main/Util.java
@@ -26,10 +26,10 @@ public class Util {
 //        return shapeArr;
 //    }
 
-    static int[][] calcBinoms(int n) {
-        int[][] binoms = new int[n + 1][];
+    static long[][] calcBinoms(int n) {
+        long[][] binoms = new long[n + 1][];
         for (int i = 0; i <= n; i++) {
-            binoms[i] = new int[i + 2];
+            binoms[i] = new long[i + 2];
             binoms[i][0] = 1;
             binoms[i][i + 1] = 0;
             for (int j = 1; j <= i; j++) {
